pyproject.toml
[tool.poetry]
name = "pet-adoption"
version = "0.1.0"
description = ""
authors = ["Ivan Berestianyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-restful = "^0.3.10"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
flask-migrate = "^4.0.7"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
